pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "vtk-python-docs"
version = "0.1.0"
description = "VTK Python documentation extraction, stub enhancement, and API generation"
readme = "README.md"
requires-python = ">=3.10"
license = {file = "LICENSE"}
authors = [
    { name = "Patrick O'Leary", email = "patrick.oleary@kitware.com" }
]
keywords = ["vtk", "documentation", "stubs", "type-hints", "pyi", "visualization"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Documentation",
    "Topic :: Software Development :: Documentation",
    "Topic :: Scientific/Engineering :: Visualization",
]
dependencies = [
    "vtk>=9.0.0",
    "typer>=0.12.0",
    "litellm>=1.0.0",
    "python-dotenv>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/patrickoleary/vtk-python-docs"
Repository = "https://github.com/patrickoleary/vtk-python-docs"
Issues = "https://github.com/patrickoleary/vtk-python-docs/issues"

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.1.0",
    "pyright>=1.1.0",
    "vulture>=2.0",
]

[project.scripts]
vtk-docs = "vtk_python_docs.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["vtk_python_docs"]

[tool.ruff]
line-length = 120
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
